refactor: add shared folder
