+++ src/shared/components/mode-toggle.tsx
@@ -3,13 +3,13 @@
 import * as React from "react";
 import { useTheme } from "next-themes";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shared/components/ui/button";
 import {
 	DropdownMenu,
 	DropdownMenuContent,
 	DropdownMenuItem,
 	DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+} from "@/shared/components/ui/dropdown-menu";
 import { CheckCheck, MoonIcon, SunIcon } from "lucide-react";
 
 export function ModeToggle() {
