+++ src/shared/components/ui/logos/github-mark.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 export const GithubMark = ({ className }: { className?: string }) => {
 	return (
