+++ src/shared/components/ui/logos/google-mark.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 export const GoogleMark = ({ className }: { className?: string }) => {
 	return (
