+++ src/shared/components/ui/spinner.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 export const Spinner = ({ className }: { className?: string }) => {
 	return (
