+++ src/shared/server/api/root.ts
@@ -1,5 +1,5 @@
-import { postRouter } from "@/server/api/routers/post"
-import { createTRPCRouter } from "@/server/api/trpc"
+import { postRouter } from "@/shared/server/api/routers/post";
+import { createTRPCRouter } from "@/shared/server/api/trpc";
 
 /**
  * This is the primary router for your server.
@@ -7,8 +7,8 @@ import { createTRPCRouter } from "@/server/api/trpc"
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-    post: postRouter,
-})
+	post: postRouter,
+});
 
 // export type definition of API
-export type AppRouter = typeof appRouter
+export type AppRouter = typeof appRouter;
