+++ src/shared/styles/globals.css
