+++ src/shared/trpc/react.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { createTRPCReact } from "@trpc/react-query";
+
+import { type AppRouter } from "@/shared/server/api/root";
+
+export const api = createTRPCReact<AppRouter>();
